+++ assets/dev/javascript/helper.js
@@ -68,7 +68,6 @@ wps_js.formatNumber = function (num, fixed = 0) {
 wps_js.select2 = function () {
     jQuery("select[data-type-show=select2]").select2();
 }
-
 const wpsSelect2 = jQuery('.wps-select2');
 const wpsBody = jQuery('body');
 const wpsDropdown = jQuery('.wps-dropdown');
@@ -161,7 +160,6 @@ if (wpsSelect2.length) {
         wpsFilterVisitor.on('click', () => wpsSelect2.select2('open'));
     }
 }
-
 /**
  * Set Tooltip
  */
@@ -454,3 +452,7 @@ jQuery(document).ready(function () {
     }
 
 });
+
+window.renderFormatNum = function (data) {
+    return wps_js.formatNumber(data)
+}