+++ src/Utils/Query.php
@@ -84,7 +84,7 @@ class Query
                 $this->setClauses[]      = "$column = %d";
                 $this->valuesToPrepare[] = $value;
             } else if (is_null($value)) {
-                $this->setClauses[] = $column = NULL;
+                $this->setClauses[] = "$column = NULL";
             }
         }
 
@@ -158,7 +158,10 @@ class Query
         }
 
         if (!empty($from) && !empty($to)) {
-            $condition                  = "DATE($field) BETWEEN %s AND %s";
+            if (strlen($from) === 10) $from .= ' 00:00:00';
+            if (strlen($to) === 10) $to .= ' 23:59:59';
+
+            $condition                  = "$field BETWEEN %s AND %s";
             $this->whereClauses[]       = $condition;
             $this->valuesToPrepare[]    = $from;
             $this->valuesToPrepare[]    = $to;
@@ -725,9 +728,12 @@ class Query
         return $this;
     }
 
-    public function removeTablePrefix($query)
+    public function removeTablePrefix($table)
     {
-        return str_replace([$this->db->prefix, 'statistics_'], '', $query);
+        $prefixLength   = strlen($this->db->prefix);
+        $table          = substr($table, $prefixLength);
+
+        return str_replace('statistics_', '', $table);
     }
 
     /**